+++ CHANGELOG.md
@@ -38,6 +38,7 @@ MarkLogic:
     debugger.
 1.  Better report `SEC-PRIV` errors when adding a query processor.
 1.  Fix the URI of the `xqterr` and `err` namespaces.
+1.  Auto-complete support for `cts:highlight` variables.
 
 ## 2020